ADD ADDITEM FUNCTIONALITY
