+++ src/components/TodoComponents/Todo.js
@@ -1,0 +1,24 @@
+const todos = [
+    {
+    task: 'Apples',
+    id: 1528817077286,
+    completed: false
+    },
+    {
+    task: 'Oranges',
+    id: 1528817084358,
+    completed: false
+    },
+    {
+    task: 'Milk',
+    id: 1528817084358,
+    completed: false
+    },
+    {
+    task: 'Toilet Paper',
+    id: 1528817084358,
+    completed: false
+    }
+]
+
+export default todos;
